refactor: remove goUsageTracker/streaming barrels, import canonical paths
The split is complete — update every importer to reference the canonical
modules directly and delete the two compat barrels:
- src/streaming.ts -> src/transports/* + src/core/transport.ts
- src/goUsageTracker.ts -> src/usage/{tracker,history,pricing,formatting}.ts

No behavior change; compile + 291 tests + lint green.
