--- src/goUsageTracker.ts
@@ -1,32 +1,0 @@
-/**
- * Barrel — the Go usage tracker was split into `src/usage/` (tracker,
- * history, pricing, formatting, usage, usageProfile, goUsageSync). This
- * module re-exports the historical public API so existing importers keep
- * working during the refactor.
- */
-export { GO_LIMITS } from "./config";
-export { GO_VENDOR } from "./providerTypes";
-export { estimateCost, type CostResolver } from "./usage/pricing";
-export {
-  GoUsageTracker,
-  startOfLocalDay,
-  normalizeCwd,
-  isCwdInWorkspace,
-  type GoUsageTrackerOptions,
-  type UsageBaselineTargets,
-  type UsageLogEntry,
-  type SessionCostSummary,
-  type PeriodUsage,
-  type UsageSummary,
-} from "./usage/tracker";
-export {
-  buildUsageSeries,
-  setHistoryReadDiagnostic,
-  sumDailyUsage,
-  type HistoryRow,
-  type ModelDayUsage,
-  type UsageDaily,
-  type UsageDayPoint,
-  type UsageSeries,
-} from "./usage/history";
-export { buildUsageQuickPickItems, formatGoUsageStatusBarText } from "./usage/formatting";
